Remember now why calling is of the form obj(methodName)(arg1, arg2) now:
It's because I want to be able to explicitly pass `self` when calling a
method, so when a method is called on a super object, its `self` is the
original receiver of the message, so it calls the correct methods. This
is shown in animalMethods#typeAndName, which calls the introduce method
of the dog.